manage focus in html example
